+++ src/lib/persistence/fcn/index.ts
@@ -1,0 +1,3 @@
+export * from "@/src/lib/persistence/fcn/fcn-persistence-repository";
+export * from "@/src/lib/persistence/fcn/fcn-persistence-service";
+export * from "@/src/lib/persistence/fcn/fcn-persistence-types";